support for characters endpoint
there is some undocumented funcionality in the API
that can be used with some reverse engineering. By
appending /characters to an anime url one gets back
the information about the character of that show.
Unfortunately this doesn't work for mangas.

Thanks to Gruzin from the API discord server for
informing me of this.