implement __getitem__ on all iterables
i still don't know if making a custom iterable for
each thing is the best way to go, i could probably
factor it out as a base class and have other inherit
it. Otherwise i could simply expose the list directly
in order to keep the other features that are needed
(e.g. pagination, url, etc)